Refactor ProjectTree and ResourcesTree components: Update ProjectTree structure to include resourcesTree and environmentsList, enhance resource management with new buildResourcesTree utility, and introduce useProjectsTrees hook for improved project data handling.

+++ view/desktop/src/domains/resource/resourceService.ts
@@ -57,7 +57,9 @@ export const resourceService: IResourceService = {
     const output = await resourceIpc.delete(projectId, input);
 
     //TODO should delete all nested resources summaries too
-    resourceSummariesCollection.delete(input.id);
+    if (resourceSummariesCollection.has(input.id)) {
+      resourceSummariesCollection.delete(input.id);
+    }
 
     return output;
   },
